requirements.txt
PyGObject>=3.42.0
requests>=2.28.0
Pillow>=9.0.0
pycairo>=1.20.0
dbus-python>=1.2.0
mutagen>=1.45.0
yt-dlp>=2023.0.0
